IOT/requirements.txt
speech_recognition
RPi.GPIO
pyaudio
gtts
pygame
torch
numpy 
librosa
soundfile
python-dotenv